+++ uliweb/core/commands.py
@@ -133,12 +133,13 @@ class Command(object):
         return get_apps(global_options.apps_dir, include_apps=include_apps, 
             settings_file=global_options.settings, local_settings_file=global_options.local_settings)
     
-    def get_application(self, global_options):
+    def get_application(self, global_options, default_settings=None):
         from uliweb.manage import make_simple_application
-        
-        return make_simple_application(project_dir=global_options.project, 
+
+        return make_simple_application(project_dir=global_options.project,
             settings_file=global_options.settings, 
-            local_settings_file=global_options.local_settings
+            local_settings_file=global_options.local_settings,
+            default_settings=default_settings
             )
         
     def run_from_argv(self, prog, subcommand, global_options, argv):