+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Uliweb Change Log
 -----------------
 * Refactor template with tornado
 * Fix Pickle default value process bug, will not convert empty value to ''
+* Add staticize command
 
 0.3.1 Version
 -----------------